# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "infinitas hosted registry",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/skills/{skill_id}/changesets": {
      "post": {
        "tags": [
          "authoring-collaboration"
        ],
        "summary": "Create Change Set",
        "operationId": "create_change_set_api_v1_skills__skill_id__changesets_post",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSetCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetView"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill or ChangeSet not found"
          },
          "409": {
            "description": "ChangeSet state or latest-version conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      },
      "get": {
        "tags": [
          "authoring-collaboration"
        ],
        "summary": "List Change Sets",
        "operationId": "list_change_sets_api_v1_skills__skill_id__changesets_get",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChangeSetView"
                  },
                  "title": "Response List Change Sets Api V1 Skills  Skill Id  Changesets Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill or ChangeSet not found"
          },
          "409": {
            "description": "ChangeSet state or latest-version conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/skills/{skill_id}/changesets/{change_set_id}": {
      "get": {
        "tags": [
          "authoring-collaboration"
        ],
        "summary": "Get Change Set",
        "operationId": "get_change_set_api_v1_skills__skill_id__changesets__change_set_id__get",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "change_set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^chg_[A-Za-z0-9_-]+$",
              "title": "Change Set Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetView"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill or ChangeSet not found"
          },
          "409": {
            "description": "ChangeSet state or latest-version conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/skills/{skill_id}/changesets/{change_set_id}/submit": {
      "post": {
        "tags": [
          "authoring-collaboration"
        ],
        "summary": "Submit Change Set",
        "operationId": "submit_change_set_api_v1_skills__skill_id__changesets__change_set_id__submit_post",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "change_set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^chg_[A-Za-z0-9_-]+$",
              "title": "Change Set Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetView"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill or ChangeSet not found"
          },
          "409": {
            "description": "ChangeSet state or latest-version conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/skills/{skill_id}/changesets/{change_set_id}/reject": {
      "post": {
        "tags": [
          "authoring-collaboration"
        ],
        "summary": "Reject Change Set",
        "operationId": "reject_change_set_api_v1_skills__skill_id__changesets__change_set_id__reject_post",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "change_set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^chg_[A-Za-z0-9_-]+$",
              "title": "Change Set Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetView"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill or ChangeSet not found"
          },
          "409": {
            "description": "ChangeSet state or latest-version conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/skills/{skill_id}/changesets/{change_set_id}/accept": {
      "post": {
        "tags": [
          "authoring-collaboration"
        ],
        "summary": "Accept Change Set",
        "operationId": "accept_change_set_api_v1_skills__skill_id__changesets__change_set_id__accept_post",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "change_set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^chg_[A-Za-z0-9_-]+$",
              "title": "Change Set Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSetAcceptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetAcceptView"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill or ChangeSet not found"
          },
          "409": {
            "description": "ChangeSet state or latest-version conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeSetAcceptRequest": {
        "properties": {
          "expected_latest_digest": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^sha256:[0-9a-f]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Latest Digest"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_latest_digest"
        ],
        "title": "ChangeSetAcceptRequest"
      },
      "ChangeSetAcceptView": {
        "properties": {
          "change_set": {
            "$ref": "#/components/schemas/ChangeSetView"
          },
          "version_id": {
            "type": "integer",
            "title": "Version Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "content_digest": {
            "type": "string",
            "title": "Content Digest"
          }
        },
        "type": "object",
        "required": [
          "change_set",
          "version_id",
          "version",
          "content_digest"
        ],
        "title": "ChangeSetAcceptView"
      },
      "ChangeSetCreateRequest": {
        "properties": {
          "base_version_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version Id"
          },
          "content_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^cnt_[A-Za-z0-9_-]+$",
            "title": "Content Id"
          },
          "proposed_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-[0-9A-Za-z.-]+)?(?:\\+[0-9A-Za-z.-]+)?$",
            "title": "Proposed Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "base_version_id",
          "content_id",
          "proposed_version"
        ],
        "title": "ChangeSetCreateRequest"
      },
      "ChangeSetView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "base_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version Id"
          },
          "candidate_content_id": {
            "type": "string",
            "title": "Candidate Content Id"
          },
          "proposed_version": {
            "type": "string",
            "title": "Proposed Version"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "submitted",
              "accepted",
              "superseded",
              "rejected"
            ],
            "title": "State"
          },
          "created_by_principal_id": {
            "type": "integer",
            "title": "Created By Principal Id"
          },
          "actor": {
            "additionalProperties": true,
            "type": "object",
            "title": "Actor"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill_id",
          "base_version_id",
          "candidate_content_id",
          "proposed_version",
          "state",
          "created_by_principal_id",
          "created_at",
          "submitted_at",
          "decided_at"
        ],
        "title": "ChangeSetView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "infinitas_auth_token"
      }
    }
  }
}